tuna: toro(plus): Parcel ABI fix.
This could probably be applied to maguro as well, but only toros crashed.

NOTE: This doesn't adjust the stack pointer in *all* places Parcels
are made on the stack. There's a couple big functions with huge
stacks that I'm not comfortable messing with right now; this fixes
the crashing problem which is good enough for now.

Change-Id: I5dd9dccb6451682538e1456aeff133521f3de825